Refactor TBA match data models
